+++ lib/models/placemark.dart
@@ -1,4 +1,4 @@
-import 'package:geolocator/models/position.dart';
+part of geolocator;
 
 /// Contains detailed placemark information.
 class Placemark {
@@ -49,13 +49,21 @@ class Placemark {
   final Position position;
 
   /// Converts a list of [Map] instances to a list of [Placemark] instances.
-  static List<dynamic> fromMaps(dynamic message) {
-    var list = message.map<Placemark>(fromMap).toList();
+  static List<Placemark> _fromMaps(dynamic message) {
+    if (message == null) {
+      throw new ArgumentError("The parameter 'message' should not be null.");
+    }
+
+    List<Placemark> list = message.map<Placemark>(_fromMap).toList();
     return list;
   }
 
   /// Converts the supplied [Map] to an instance of the [Placemark] class.
-  static Placemark fromMap(dynamic message) {
+  static Placemark _fromMap(dynamic message) {
+    if (message == null) {
+      throw new ArgumentError("The parameter 'message' should not be null.");
+    }
+
     final Map<dynamic, dynamic> placemarkMap = message;
 
     return new Placemark._(
@@ -69,7 +77,7 @@ class Placemark {
       subLocality: placemarkMap["subLocality"] ?? "",
       thoroughfare: placemarkMap["thoroughfare"] ?? "",
       subThoroughfare: placemarkMap["subThoroughfare"] ?? "",
-      position: Position.fromMap(placemarkMap['location']),
+      position: Position._fromMap(placemarkMap['location']),
     );
   }
 }